[refactor]: Ajustes de componentizacao dos detalhes de renda
