+++ src/components/Card/styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import colors from '../../utils/colors';
+import colors from '../../constants/colors';
 
 export const Container = styled.div`
   display: flex;
